pyproject.toml
[project]
name = "pkg-info-validator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "imageio>=2.37.0",
    "jsonschema>=4.23.0",
    "nuitka==2.6.4",
    "pyinstaller>=6.11.1",
    "pywebview>=5.4",
    "pyyaml>=6.0.2",
    "qtpy>=2.4.2",
]
